artifacts: don't assign chain as an attribute
We only used it to resolve relative paths to chain.cwd, but that should
always be the actual cwd of the ckis process, so we can just use pathlib
to resolve relative paths to the current working directory.
Decoupling artifacts from the chain they were created by is better, as
it is not a property of the artifact. The chain already keeps the
artifacts it owns in chain.store.