prune leftover artifacts from previous run
At the end of each run, the installed artifacts are stored on disk.
After the next run finishes, this list is compared with the artifacts
that just got installed. If some did not get overwritten, we remove
them, so we don't get dangling artifacts when changing the config.

Only artifacts from the same chain and kernel version are pruned.
Pruning artifacts from other chains/kernels needs to be done centrally.
This will be added at some point in the future.